Changed method of diffing previews database.
Previously the diff for the previews database and the previews cache
only worked when previews were added to the database. This has been
fixed so a full diff of adding and removing previews is handled. May not
be the most optimal solution but this can fixed in the future if needed.

Added settings implementation but only in memory.